MLBEDSW-11315: Use correct precision for Explicit quantization
In order to match the reference we must cast scale to float
and then back to double.

+++ ethosu/regor/compiler/tflite_graph_optimiser.cpp
@@ -2770,7 +2770,8 @@ Operation *TFLiteGraphOptimiser::ConvertQuantizetoExplicit(Graph *const graph, O
 
     double ifmScale = ifmQuant.Scale().Dequantize();
     double ofmScale = ofmQuant.Scale().Dequantize();
-    QuantizedScale quantScale = double(float(ifmScale) / float(ofmScale));
+    // Cast to float and back to double is necessary to match the reference.
+    QuantizedScale quantScale = double(float(ifmScale)) / double(float(ofmScale));
 
     ofmQuant.scales.clear();
     ofmQuant.scales.push_back(quantScale);
